Reworked stop variable for continous looping
Having the user enter 0 for a continuous scan was causing serious problems with comparisons later in the program, so I changed it to be a Boolean for looping and an integer for limited runs. This should fix the bot refusing to continuously run and/or not running the specified amount of times.