Malware Flagging
This adds a rudimentary system for flagging potential malware. It identifies suspicious powershell artifacts, python syscalls and socket work, and base64 encoded files and URLs.

This system is simply a stand-in for the Yara system that it is in the works. Once yara support is added, malware detection can be customized and extended via a set of yara rules. 

This current implementation shaves off a fair amount of unnecessary code by making the archive_engine function save files during both filter and raw search modes.