Major changes to list variables
This update removes the arch_mode variable completely, as it was basically useless.

This update also fixes a mistake with how keylist, reglist, and blacklist variables were stored. Parser was fetching them as a string, causing archive_engine to loop through individual letters as opposed to elements. The solution is to store the lists as a comma delimited string under the name *str, and convert them to their *list variable counterparts on lines 400-413